fix(refactor) : Car class
- method 이름 용도에 맞게 변경
- changeGears의 매개변수 추가
